requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dropbox==9.4.0
emoji==0.5.4
future==0.18.2
idna==2.8
pycparser==2.19
python-telegram-bot==12.3.0
PyYAML==5.3
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
tornado==6.0.3
urllib3==1.25.8
